PodDataDisplay code cleaned up, functions moved to handlers
